CPs/ingesta.almacenamiento-distribuido/requirements.txt
kafka-python-ng
hdfs
requests
jsonschema
rich